Version 1.0 of Weather Console Application:
This is the initial version of the weather console app. It allows users to enter an API Key and Zipcode to then receive weather data.

Upcoming versions will have the program run in the background to give weather updates every 10-20mins and with possible integrations to a webapp using .ASPNET